Various minor physics and entity changes.
Add const to several values; avoid implicit int-to-double conversion; make sure normal vectors are normalized; revise voxelEntityMap insertion; change arbitrary distance values to constexpr locals; remove unused success boolean; compare entity distance squared instead of distance; don't pass C array to rayQuadIntersection().

+++ OpenTESArena/src/Entities/EntityManager.cpp
@@ -432,8 +432,7 @@ void EntityManager::getEntityVisibilityData(const Entity &entity, const Double2 
 		}
 	}();
 
-	outVisData.anglePercent = static_cast<float>(
-		std::clamp(animAngle / Constants::TwoPi, 0.0, Constants::JustBelowOne));
+	outVisData.anglePercent = std::clamp(animAngle / Constants::TwoPi, 0.0, Constants::JustBelowOne);
 
 	const int stateIndex = [&outVisData, stateCount]()
 	{
